// src/types/Stats.ts
export type Stats = {
    visits?: number;
    total_profit?: number;
    total_repairs?: number;
    total_orders?: number;
    total_products?: number;
    updatedAt?: number;
    createdAt?: number;
}

export interface StatsApi{
    visits?: number;
    total_profit?: number;
    total_repairs?: number;
    total_orders?: number;
    total_products?: number;
    updatedAt?: number;
    createdAt?: number;
}